# querier: fix performance when ingesters stream samples