#### What this PR does

The conversion from LabelAdapters to Labels is expensive, now the underlying data structure is not identical, so do the comparison without converting.

Code copied from old Labels implementation.

#### Checklist

- [x] Tests updated
- NA Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`