@bboreham 2023-08-25
I extended the tests to check sorting, and added a benchmark, here run with `-tags stringlabels`:

```
name                   old time/op    new time/op    delta
TimeSeriesSeriesSet-8    1.09ms ± 1%    0.12ms ± 1%   -89.22%  (p=0.002 n=6+6)

name                   old alloc/op   new alloc/op   delta
TimeSeriesSeriesSet-8    1.56MB ± 0%    0.00MB ± 0%  -100.00%  (p=0.000 n=5+6)

name                   old allocs/op  new allocs/op  delta
TimeSeriesSeriesSet-8     32.5k ± 0%      0.0k ± 0%  -100.00%  (p=0.000 n=5+6)
```

I didn't put a lot of effort into checking the comparison function since it is copy-pasted from Prometheus labels.
I can copy-paste the test if you think I should.  It's 85 lines.